Crate navigator_rs

Navigator-rs

This crate serves as the entry point for embedding applications using Rust on Blue Robotics’s Navigator.

The Navigator board has the Raspberry Pi HAT form factor, which allows you to easily attach it to a Raspberry Pi 4 board. Then you can unleash the power of Navigator to develop new solutions or interact with your ROV hardware.

The board offers the following capabilities:

Control:

LEDs

PWM (Pulse Width Modulation) with 16 channels

Measurements:

ADC (Analog Digital Converter) entries

Magnetic field

Acceleration

Angular velocity

Temperature

Pressure

Currently, it supports armv7 and aarch64 architectures, which are the official defaults for BlueOS. However, despite using the embedded-hal concept, new ports can be added as long as the platform matches the hardware design and specifications.

For more information, including installation instructions, schematics, and hardware specifications, please check the navigator hardware setup guide.

How to use this crate:

The capabilities of the Navigator can be accessed through the Navigator object. Please check the examples in the Implementations section.

Structs

AxisData

The AxisData struct encapsulates values for the x, y, and z axes.

Navigator

The Navigator struct contains various components used for navigator. It includes PWM control, pressure and temperature sensing, analog-to-digital conversion, inertial measurement unit, magnetometer, and LEDs control.

NavigatorBuilder

SensorData

Reads all sensors and stores on a single structure.

Enums

AdcChannel

Set of available options to select ADC’s channel.

NavigatorVersion

Peripherals

PiVersion

UserLed

Set of options to control navigator’s LEDs.