Struct navigator_rs::Navigator

pub struct Navigator { /* private fields */ }

The Navigator struct contains various components used for navigator. It includes PWM control, pressure and temperature sensing, analog-to-digital conversion, inertial measurement unit, magnetometer, and LEDs control.

All this components were integrated and abstracted to be used directly from navigator module.

Please check Implementations and it’s examples, then start to coding your applications!

Implementations

impl Navigator

pub fn new() -> Navigator

pub fn init(&mut self)

pub fn self_test(&mut self) -> bool

pub fn pwm_enable(&mut self, state: bool)

Sets the PWM IC to be enabled through firmware and OE_pin.

Arguments

state - The state of PWM output, it’s enabled with a true logic value.

Examples

Please check set_pwm_channel_value.

pub fn set_pwm_channel_value(&mut self, channel: PwmChannel, value: u16)

Sets the Duty Cycle (high value time) of selected channel.

On PCA9685, this function sets the OFF counter and uses ON value as 0.

Further info

Check 7.3.3 LED output and PWM control

Examples

use navigator_rs::{Navigator, PwmChannel};

let mut nav = Navigator::new();

nav.init();
nav.pwm_enable(true);

nav.set_pwm_freq_prescale(99); // sets the pwm frequency to 60 Hz
nav.set_pwm_channel_value(PwmChannel::Ch1, 2048); // sets the duty cycle to 50%

pub fn set_pwm_channels_value<const N: usize>( &mut self, channels: &[PwmChannel; N], value: u16 )

Like set_pwm_channel_value. This function sets the Duty Cycle for a list of multiple channels.

Examples

use navigator_rs::{Navigator, PwmChannel};

let mut nav = Navigator::new();

nav.init();
nav.pwm_enable(true);
nav.set_pwm_freq_prescale(99); // sets the pwm frequency to 60 Hz

let channels: [PwmChannel; 3] = [PwmChannel::Ch1, PwmChannel::Ch1, PwmChannel::Ch2];
let values: [u16; 3] = [200, 1000, 300];

nav.set_pwm_channels_value(&channels, 2048); // sets the duty cycle according to the list.

pub fn set_pwm_channels_values<const N: usize>( &mut self, channels: &[PwmChannel; N], values: &[u16; N] )

Like set_pwm_channel_value. This function sets the Duty Cycle for a list of multiple channels with multiple values.

Examples

use navigator_rs::{Navigator, PwmChannel};

let mut nav = Navigator::new();

nav.init();
nav.pwm_enable(true);
nav.set_pwm_freq_prescale(99); // sets the pwm frequency to 60 Hz

let channels: [PwmChannel; 3] = [PwmChannel::Ch1, PwmChannel::Ch1, PwmChannel::Ch2];
let values: [u16; 3] = [200, 1000, 300];

nav.set_pwm_channels_values(&channels, &values); // sets the duty cycle according to the lists.

pub fn set_pwm_freq_prescale(&mut self, value: u8)

Sets the PWM frequency of Navigator.

It changes the PRE_SCALE value on PCA9685.

The prescaler value can be calculated for an update rate using the formula:

prescale_value = round(clock_freq / (4096 * desired_freq)) - 1.

The minimum prescaler value is 3, which corresponds to 1526 Hz. The maximum prescaler value is 255, which corresponds to 24 Hz.

If you want to control a servo, set a prescaler value of 100. This will correspond to a frequency of about 60 Hz, which is the frequency at which servos work.

Internally, this function stops the oscillator and restarts it after setting the prescaler value if it was running.

Re-run the set_pwm_channel_value() is required.

Further info

Check 7.3.5 - PWM frequency PRE_SCALE

Examples

use navigator_rs::{Navigator, PwmChannel};

let mut nav = Navigator::new();

nav.init();
nav.pwm_enable(true);

nav.set_pwm_freq_prescale(99); // sets the pwm frequency to 60 Hz

nav.set_pwm_channel_value(PwmChannel::Ch1, 2048); // sets the duty cycle to 50%

pub fn set_pwm_freq_hz(&mut self, freq: f32)

Sets the pwm frequency in Hertz of Navigator.

The navigator module uses a crystal with a 24.5760 MHz clock. You can set a value for a frequency between 24 and 1526 Hz.

Examples

use navigator_rs::{Navigator, PwmChannel};
use std::thread::sleep;
use std::time::Duration;

let mut nav = Navigator::new();

nav.init();
nav.pwm_enable(true);

let mut i: f32 = 10.0;

loop {
    nav.set_pwm_freq_hz(i);
    nav.set_pwm_channel_value(PwmChannel::Ch1, 2048); // sets the duty cycle to 50%
    i = i + 10.0;
    sleep(Duration::from_millis(1000));
}

pub fn get_led(&mut self, select: UserLed) -> bool

Gets the selected navigator LED output state. The true state means the LED is on.

Arguments

select - A pin selected from UserLed.

Examples

use navigator_rs::{UserLed, Navigator};
use std::thread::sleep;
use std::time::Duration;

let mut nav = Navigator::new();

nav.init();
loop {
    let logic_value: bool = nav.get_led(UserLed::Led2);
    println!("Blue LED logic value is {logic_value}.");
    nav.set_led_toggle(UserLed::Led2);
    sleep(Duration::from_millis(1000));
}

pub fn set_led(&mut self, select: UserLed, state: bool)

Sets the selected navigator LED output.

Arguments

select - A pin selected from UserLed.

state - The value of output, LED is on with a true logic value.

Examples

use navigator_rs::{UserLed, Navigator};
use std::thread::sleep;
use std::time::Duration;

let mut nav = Navigator::new();

nav.init();
loop {
    nav.set_led(UserLed::Led1, true);
    sleep(Duration::from_millis(1000));
    nav.set_led(UserLed::Led1, false);
    sleep(Duration::from_millis(1000));
}

pub fn set_led_toggle(&mut self, select: UserLed)

Toggle the output of selected LED.

Arguments

select - A pin selected from UserLed.

Examples

use navigator_rs::{UserLed, Navigator};
use std::thread::sleep;
use std::time::Duration;

let mut nav = Navigator::new();

nav.init();
loop {
    nav.set_led_toggle(UserLed::Led1);
    sleep(Duration::from_millis(1000));
}

pub fn set_led_all(&mut self, state: bool)

Set all LEDs on desired state ( Blue, Green and Red ).

Arguments

state - The value of output, LED is on with a true logic value.

Examples

use navigator_rs::{Navigator};
use std::thread::sleep;
use std::time::Duration;

let mut nav = Navigator::new();

nav.init();
loop {
    nav.set_led_all(true);
    sleep(Duration::from_millis(1000));
    nav.set_led_all(false);
    sleep(Duration::from_millis(1000));
}

pub fn set_neopixel(&mut self, array: &[[u8; 3]])

Set the values of the neopixel LED array.

Arguments

array - A 2D array containing RGB values for each LED. Each inner array is a [u8; 3] representing the Red, Green and Blue from a LED.

Example

use navigator_rs::{Navigator};

let mut nav = Navigator::new();

nav.init();
let mut leds = [[0, 0, 255], [0, 255, 0], [255, 0, 0]];
nav.set_neopixel(&mut leds);

This will set the first LED to blue, second to green, and third to red.

pub fn read_mag(&mut self) -> AxisData

Reads the magnetometer Ak09915 of Navigator.

Measurements in [µT]

Examples

use navigator_rs::{Navigator};
use std::thread::sleep;
use std::time::Duration;

let mut nav = Navigator::new();
nav.init();

loop {
    let mag = nav.read_mag();
    println!("mag values: X={}, Y={}, Z={} [µT]", mag.x, mag.y, mag.z);
    sleep(Duration::from_millis(1000));
}

pub fn read_temperature(&mut self) -> f32

Reads the temperature using BMP280 of Navigator.

Measurements in [˚C]

Examples

use navigator_rs::{Navigator};
use std::thread::sleep;
use std::time::Duration;

let mut nav = Navigator::new();
nav.init();

loop {
    let temperature = nav.read_temperature();
    println!("temperature value: {temperature} [˚C]");
    sleep(Duration::from_millis(1000));
}

pub fn read_altitude(&mut self) -> f32

Reads the altitude based on pressure values measured by BMP280 of Navigator.

Measurements in [m]

Examples

use navigator_rs::{Navigator};
use std::thread::sleep;
use std::time::Duration;

let mut nav = Navigator::new();
nav.init();

loop {
    let altitude = nav.read_altitude();
    println!("altitude value: {altitude} [m]");
    sleep(Duration::from_millis(1000));
}

pub fn read_pressure(&mut self) -> f32

Reads the pressure based on BMP280 of Navigator.

Measurements in [kPa]

Examples

use navigator_rs::{Navigator};
use std::thread::sleep;
use std::time::Duration;

let mut nav = Navigator::new();
nav.init();

loop {
    let pressure = nav.read_pressure();
    println!("pressure value: {pressure} [kPa]");
    sleep(Duration::from_millis(1000));
}

pub fn read_adc_all(&mut self) -> ADCData

Reads the ADC based on ADS1115 of Navigator.

Measurements in [V]

Same as read_adc, but it returns an array with all channel readings

Examples

use navigator_rs::{Navigator};
use std::thread::sleep;
use std::time::Duration;

let mut nav = Navigator::new();
nav.init();

loop {
    let adc_data = nav.read_adc_all();
    println!(
        "adc values per channel: 1: {}, 2: {}, 3: {}, 4: {}",
        adc_data.channel[0], adc_data.channel[1], adc_data.channel[2], adc_data.channel[3]
    );
    sleep(Duration::from_millis(1000));
}

pub fn read_adc(&mut self, channel: AdcChannel) -> f32

Reads the ADC based on ADS1115 of Navigator.

Measurements in [V]

Examples

use navigator_rs::{AdcChannel, Navigator};
use std::thread::sleep;
use std::time::Duration;

let mut nav = Navigator::new();
nav.init();

loop {
    println!("ADC channel 1 value: {} [V]", nav.read_adc(AdcChannel::Ch1));
    sleep(Duration::from_millis(1000));
}

pub fn read_accel(&mut self) -> AxisData

Reads acceleration based on ICM20689 of Navigator.

Measurements in [m/s²]

Examples

use navigator_rs::{Navigator};
use std::thread::sleep;
use std::time::Duration;

let mut nav = Navigator::new();
nav.init();

loop {
    let accel = nav.read_accel();
    println!("accel values: X={}, Y={}, Z={} [m/s²]", accel.x, accel.y, accel.z);
    sleep(Duration::from_millis(1000));
}

pub fn read_gyro(&mut self) -> AxisData

Reads angular velocity based on ICM20689 of Navigator.

Measurements in [rad/s]

Examples

use navigator_rs::{Navigator};
use std::thread::sleep;
use std::time::Duration;

let mut nav = Navigator::new();
nav.init();

loop {
    let gyro = nav.read_gyro();
    println!("accel values: X={}, Y={}, Z={} [rad/s]", gyro.x, gyro.y, gyro.z);
    sleep(Duration::from_millis(1000));
}

pub fn read_all(&mut self) -> SensorData

Reads all sensors and stores on a single structure.

Examples

use navigator_rs::{Navigator, SensorData};
use std::thread::sleep;
use std::time::Duration;

let mut nav = Navigator::new();
nav.init();

loop {
    println!("---------------reading---------------");
    let sensors_data = nav.read_all();
    println!("{sensors_data:#?}");
    sleep(Duration::from_millis(1000));
}

pub fn fmt_debug(&mut self) -> impl Debug