Crate navigator_rs

Navigator-rs

This crate serves as the entry point for embedding applications using Rust on Blue Robotics’s Navigator.

The Navigator board has the Raspberry Pi HAT form factor, which allows you to easily attach it to a Raspberry Pi 4 board. Then you can unleash the power of Navigator to develop new solutions or interact with your ROV hardware.

The board offers the following capabilities:

Control:

LEDs

PWM (Pulse Width Modulation) with 16 channels

Measurements:

ADC (Analog Digital Converter) entries

Magnetic field

Acceleration

Angular velocity

Temperature

Pressure

Currently, it supports armv7 and aarch64 architectures, which are the official defaults for BlueOS. However, despite using the embedded-hal concept, new ports can be added as long as the platform matches the hardware design and specifications.

For more information, including installation instructions, schematics, and hardware specifications, please check the navigator hardware setup guide.

How to use this crate:

The capabilities of the Navigator can be accessed through the Navigator object. Please check the examples in the Implementations section.

Structs

ADCData

Encapsulates the value of ADC’s four channels.

AxisData

The AxisData struct encapsulate values for the x, y, and z axes.

Led

The Led struct represents the 3 LEDs on navigator board.

Navigator

The Navigator struct contains various components used for navigator. It includes PWM control, pressure and temperature sensing, analog-to-digital conversion, inertial measurement unit, magnetometer, and LEDs control.

NavigatorBuilder

Build pattern structure

Pwm

The Pwm struct represents a PWM (Pulse Width Modulation) controller with a PCA9685 chip and it’s output enable pin.

SensorData

Encapsulates the value of all sensors on the board.

Enums

AdcChannel

Set of available options to select ADC’s channel.

PwmChannel

Set of available options to select PWM’s channel.

UserLed

Set of options to control navigator’s LEDs.