Struct navigator_rs::Navigator

pub struct Navigator { /* private fields */ }

Implementations

impl Navigator

pub fn new() -> Navigator

pub fn init(&mut self)

pub fn self_test(&mut self) -> bool

pub fn pwm_enable(&mut self)

pub fn pwm_disable(&mut self)

pub fn set_pwm_channel_value(&mut self, channel: PwmChannel, value: u16)

Sets the Duty Cycle (high value time) of selected channel.

On PCA9685, this function sets the OFF counter and uses ON value as 0.

Further info

Check 7.3.3 LED output and PWM control

Examples

use navigator_rs::{Navigator, PwmChannel};

let mut nav = Navigator::new();

nav.init();
nav.pwm_enable();

nav.set_pwm_freq_prescale(99); // sets the pwm frequency to 60 Hz
nav.set_pwm_channel_value(PwmChannel::Ch0, 2048); // sets the duty cycle to 50%

pub fn set_pwm_channels_value<const N: usize>( &mut self, channels: &[PwmChannel; N], value: u16 )

Like set_pwm_channel_value. This function sets the Duty Cycle for a list of multiple channels.

Examples

use navigator_rs::{Navigator, PwmChannel};

let mut nav = Navigator::new();

nav.init();
nav.pwm_enable();
nav.set_pwm_freq_prescale(99); // sets the pwm frequency to 60 Hz

let channels: [PwmChannel; 3] = [PwmChannel::Ch0, PwmChannel::Ch1, PwmChannel::Ch2];
let values: [u16; 3] = [200, 1000, 300];

nav.set_pwm_channels_value(&channels, 2048); // sets the duty cycle according to the list.

pub fn set_pwm_channels_values<const N: usize>( &mut self, channels: &[PwmChannel; N], values: &[u16; N] )

Like set_pwm_channel_value. This function sets the Duty Cycle for a list of multiple channels with multiple values.

Examples

use navigator_rs::{Navigator, PwmChannel};

let mut nav = Navigator::new();

nav.init();
nav.pwm_enable();
nav.set_pwm_freq_prescale(99); // sets the pwm frequency to 60 Hz

let channels: [PwmChannel; 3] = [PwmChannel::Ch0, PwmChannel::Ch1, PwmChannel::Ch2];
let values: [u16; 3] = [200, 1000, 300];

nav.set_pwm_channels_values(&channels, &values); // sets the duty cycle according to the lists.

pub fn set_pwm_freq_prescale(&mut self, value: u8)

Sets the PWM frequency of Navigator.

It changes the PRE_SCALE value on PCA9685.

The prescaler value can be calculated for an update rate using the formula:

prescale_value = round(clock_freq / (4096 * desired_freq)) - 1.

The minimum prescaler value is 3, which corresponds to 1526 Hz. The maximum prescaler value is 255, which corresponds to 24 Hz.

If you want to control a servo, set a prescaler value of 100. This will correspond to a frequency of about 60 Hz, which is the frequency at which servos work.

Internally, this function stops the oscillator and restarts it after setting the prescaler value if it was running.

Re-run the set_pwm_channel_value() is required.

Further info

Check 7.3.5 - PWM frequency PRE_SCALE

Examples

use navigator_rs::{Navigator, PwmChannel};

let mut nav = Navigator::new();

nav.init();
nav.pwm_enable();

nav.set_pwm_freq_prescale(99); // sets the pwm frequency to 60 Hz

nav.set_pwm_channel_value(PwmChannel::Ch0, 2048); // sets the duty cycle to 50%

pub fn set_pwm_freq_hz(&mut self, freq: f32)

Sets the pwm frequency in Hertz of Navigator.

The navigator module uses a crystal with a 24.5760 MHz clock. You can set a value for a frequency between 24 and 1526 Hz.

Examples

use navigator_rs::{Navigator, PwmChannel};
use std::thread::sleep;
use std::time::Duration;

let mut nav = Navigator::new();

nav.init();
nav.pwm_enable();

let mut i: f32 = 10.0;

loop {
    nav.set_pwm_freq_hz(i);
    nav.set_pwm_channel_value(PwmChannel::Ch0, 2048); // sets the duty cycle to 50%
    i = i + 10.0;
    sleep(Duration::from_millis(1000));
}

pub fn set_pwm_off(&mut self)

pub fn set_pwm_on(&mut self)

pub fn set_led_on(&mut self)

pub fn set_led_off(&mut self)

pub fn read_mag(&mut self) -> AxisData

pub fn read_temperature(&mut self) -> f32

pub fn read_altitude(&mut self) -> f32

pub fn read_pressure(&mut self) -> f32

pub fn read_adc_all(&mut self) -> ADCData

pub fn read_adc(&mut self, channel: AdcChannel) -> i16

pub fn read_accel(&mut self) -> AxisData

pub fn read_gyro(&mut self) -> AxisData

pub fn read_all(&mut self) -> SensorData

pub fn fmt_debug(&mut self) -> impl Debug